Remove auto-maximize on capture start and end
The profiler is the only tool window in Android Studio
that exhibits auto-expanding behavior on use. In practice,
this is seen as when the user starts and/or ends a recording,
the profiler window height will go to the max window size.
This UI inconsistency has caused some consequent UI/UX issues.

One example of such problems is on finishing a recording
of a trace, the profiler will auto-expand to max height.
So, when the user jumps to source via a callstack sample,
they will not see that they navigated to a new file as
the profiler covers the entire screen.

Thus, this change removes this auto-maximizing
functionality, leaving the window resizing completely up
to the user. This tool window behavior is now consistent
with the other tool windows.

Fixes: 65623636
Test: updated
Change-Id: Ibd7a5fa8d1d9015a0743b2885716fb767d946e8b