Only use error panel size if it is going to be displayed
When sizing the SceneViewPanel, the error panel should only be taken
into account if it is actually going to be rendered. For example, when
rendering a View-based layout, the layout should be displayed always
whether it has an error or not. So in that case we should keep using the
layout size and not the error panel size.

Fixes: 232191108
Test: checked manually
Change-Id: I2cddde8d09a7f81a881fc390cbeb2db65b8be31d