Fix no Layout Validation tab when reopening the project
There is no persist order between opening editors and initializing
tool windows when a project is reopened.
In the below scenario, the ToolWindow of Layout Validation Tool is not
available even when a Layout Editor is selected:

Assuming a Layout Editor is opened before closing the project. In this
case, IJ will automatically open that Layout Editor when reopening the
project.

If the initializing follows the below order:

1. Opening the project.
2. (Layout Editor is not opened yet)
3. VisualizationToolWindowFactor.shouldBeAvailable() is called and it
   returns False. This results Validation Tool is not available.
4. IJ opens Layout Editor automatically. At this moment, there is no
   FileManagerListener to make Validation Tool becomes available.
   Validation Tool is still not available.
5. VisualizationToolWindowFactory.init() is called and the
   FileManagerListener is registered. But since the Layout Editor is
   opened before listener is registered, the Validation Tool is still
   not available.

Result: A Layout Editor is opened, but Validation Tool is not
        available.

To resolve the issue, add an one-shot listener to update the available
state after Validation Tool is registered completely.

Note: Changing ToolWindow.avaialbe is not allowed by IJ's platform and
it results a IllegalStateException. Thus we have to use an one-shot
listener to change the availability.

Fixes: 247493277
Test: manually and add the regression test.
Change-Id: I29991f31f1632c68081e69f4cb9e7887412e3d45