Use tag name to display TextViews and Buttons in Blueprint
Blueprint mode cannot support all the possible ways to style texts
inside buttons and text views. Currently it still displays the full text
of the component, which may lead to confusion for developers who see the
correct text with the wrong styling.
Instead, this simply displays the tag name of the component in the
center of the view.

Fixes: 123603252
Test: tests updated
Change-Id: I31f577a9b973344f4696b329e5615a04113e26d6