[Sync] Be more careful generating AGP version strings
There are mildly strong assumptions in various places, shortly to be
made explicit, about the format of an AGP version string -- beyond the
conventions of maven / gradle coordinates.  In the injected model
builder, be careful to respect those conventions, including providing
an explicit micro version and zero-padding the preview version number
if any.

There exist historical (pre-3.1.x) versions of AGP which did not
zero-pad the preview version number: this code to generate a version
string from the parsed version numbers will not be called when dealing
with any such version.

Bug: 242691473
Test: existing
Change-Id: Ib1b5bd314f3bff45fd68370d0d1d13b0143135f0