Fix the library version not found detection
The original fix only works when a dev jar is used.
When dev jar is not used the problem will be detected earlier when
we try to determine the version of compose:ui:ui.
Add the banner here too.

The tests all use a dev jar so we don't have tests for this case.

Bug: 246355403
Test: Manual
Change-Id: Ic8328065af7e870379c40f8fc389749f9f3c2d68