[AUA] Remove upgrade-to-latest fix from Old-AGP Sync issue
If the AGP version is too old to be supported, it is extremely
unlikely that an automated attempt to upgrade to the latest AGP
version known to Studio will be successful.  It's much more likely
that guiding the user to upgrade to the minimum supported version will
lead to a good outcome.  Also change a related text label to not
promise upgrading to the latest version.

Bug: 183722697
Test: existing test adjusted
Change-Id: I2ccf3fecdc0c063a106fed2207ae4f14d57df139