Re-enable LiteralsTest
This change addresses the small flakiness caused by
ProcessCanceledException. First, it removes the catch from the non
blocking actions to allow the action to retry when possible. Second, it
introduces a retry for tests to avoid on cancellation failing the test.

Also adding additional logging to help debugging the problem in
LiveLiteralsServiceTest.

Test: LiteralsTest, LiveLiteralsServiceTest
Bug: 253897699
Bug: 194631917
Change-Id: Iaf61cb61ef4126c8753a3bd85492e3cdbc6740c4