Add ability to target specific Emulator with ADB commands
Right now ADB commands are run without specifying a device. For
many commands this works fine, but there are some situations and
commands where providing an explicit device may be needed.

For instance, any test with more than one emulator running will
need this ability. Any test that wants to be robust to running on
a machine with a device connected to it will also want to be
explicit about what device should receive commands.

Bug: None
Test: None
Change-Id: Ib517af26ad34a4ec1ae0c26bb9ab5161cb0d7a6e