Don't build tests in MakeGradleProjectAction
This moves it back to the behaviour we had in D.
A bug was fixed in the GradleBuildExecutor which was preventing tests
from being build when TestCompileType.ALL was passed.

This change switches to TestCompileType.NONE which keeps the behaviour
of this action what it was pre-bug fix.

Bug: N/A
Test: updated
Change-Id: I8274b4c01a13185ba13507af68bcd53f36494d15
(cherry picked from commit 48598f80169e15e1e2c84a074fbca8e4c297cc4e)