Make conversion b/w trace configurations direct
Before this change, the conversion between ProfilingConfiguration
and CpuProfilerConfig was done by converting the source config to
a UserOptions proto, and then converting the UserOptions proto into
the destination config. This indirection has been removed by
creating methods that allow direct conversion from CpuProfilerConfig
to ProfilerConfiguration and vice-versa.

This change is part of a larger effort to remove the use of the
UserOptions proto. To do this, we first will look to eliminate
the usage of fromProto and toProto methods defined in both the
CpuProfilerConfig and ProfilerConfiguration classes, which are
both major consumers of the UserOptions proto. Creating a direct
conversion between these objects removes one large use case for
these UserOptions-reliant methods.

Bug: 258515754
Test: added
Change-Id: I548e99bea6770baef9b6b7ef16c1eb2d0be9bf1f