Update projectJdk after a successful gradle sync
In order to avoid unresolved symbols in the project due to corrupted
roots containing wrong paths, the projectJdk will be updated after
a successful sync only when all modules in the project belong to
Gradle root and to the same Gradle root.

In addition, we are using suggestSdkName to determine the jdk.table
entry given a jdkPath, validating its roots and recreate it if
necessary before assigning it to projectJdk.

As a note we avoid picking the first valid JDK entry given a
jdk path given that the jdk.table isn't shared cross users
creating conflicts in misc.xml if they have it on VCS.

Different integration tests were added to validate the expected
behaviour covering different project JDK scenarios

Bug: 236692992
Test: GradleSyncUpdatesProjectJdkIntegrationTest
Change-Id: I2ceb59e827ee86fc174fc954c9778e7bdfe913cf