Allow namespace to differ from packageName
This is primarily needed for tests to simulate the case where we create
a new activity in a subpackage of an existing module. (If we're creating
a new module, the activity will always be in the module's namespace
package; we don't provide separate inputs for those.)

Add a test demonstrating that doing so breaks the Empty Activity
template (in two ways: failing to import the R class, and failing to
appropriately qualify the activity in AndroidManifest.xml).

Replace withNewPackage customizer in TemplateTest: its work is undone by
ProjectChecker.createProject, and we also need to update the directories
from the new package using setModuleRoots.

Bug: 257290395
Bug: 255821913
Test: Included
Change-Id: I36332453abcfe6dd948f72d33f770c7788d38071