Optimize package name query
Cache information on package names and AndroidFacets.
This code is on the hot path, invoked by the compiler
front end 1000s of times. Previously, every invocation
would allocate an array that could be 1000s of elements
for large projects, and facet iteration was inefficient.

This change adds a project-wide cache that store mapping
from package name to AndroidFacets with such name, and it
also stores scopes that should be used for the manifest
index when querying for the package name.

Performance impact is that refactoring for a very large
project (5500 Gradle subprojects) finishes in ~15seconds,
while previously it would cause multi-minute freezes.
Also, the code analysis finishes in ~10-15seconds
for some files, while it used to take 40+seconds.

Bug: 248576926
Test: manually tested against 5500 Gradle subproject
      Android build
Change-Id: I21a5d9d46df6344ccc0c9492287a861dd20dcd8d
(cherry picked from commit efe443572acad5ffdbfbf9eb3124f8d3583d07ed)