Add time-based retry to LightClassesTestBase
When these tests are failing occasionally, it's due to an element not
being found under the caret. Nothing in the logs indicates any errors
that I can find. It's unclear whether this happens because something's
gone wrong with reference resolution, or whether we just needed to wait.

Busy waiting is clearly sub-optimal, but will help determine which
failure case we're dealing with. If it does turn out that we just need
to wait, we can investigate futher whether there's a better trigger to
use for continue the test than busy looping.

Bug: 256845633
Test: LightClassesTests
Change-Id: I7b01e0151b7d1bfe480ff6437841207873f84af1