Replace more UserOptions.TraceType with enum
In a previous change (ag/20427704), the
UserOptions.TraceType has been replaced with a studio-side
enum. However, there were still some instances of the
UserOptions.TraceType being used. These instances were
less trivial to replace, so they were saved for another
change. This is because they required changes to method
parameters and manipulation of TraceConfiguration creation.

With that said, there are still a few instances of UserOptions's
TraceType still present even after this change. But they are either
there to maintain backwards-compatibility or to keep old tests
from failing.

Bug: 258515754
Test: updated
Change-Id: Ibd23cbb697b1f79851d2a1d1229ac0160d4d3602