Remove preview pinning legacy code
This change removes the Compose Preview pinning flags and related code.
The flags have been disabled for a while now, and the future plans for
the Compose Preview UX indicates we're not going to support pinning as
it was originally implemented. It's much better for us to clean up the
legacy code related to pinning and re-write the feature from scratch
when the time comes.

Test: compose-designer tests should still pass
Bug: N/A
Change-Id: I5d6188200b6fb3dd850d9a4999eac6d951757e4b