Fix getting application ID of test only module
The application id of a test only module is retrieved from the module
the test only module targets. However, the build variant that is
targeted can be different from the build variant selected in the IDE.

The IDE does not not receive the full details of all variants as a
result of syncing a project but it receives some basic info that
includes application ids.

The existing `GradleApplicationIdProvider` uses the full variant data to
fetch application ids and thus fails on test only modules targeting
other build variants.

Split instantiation of `GradleApplicationIdProvider` into two modes:
(1) the default mode mode, in which it is supposed to work on all
modules and (2) a special mode for base modules which is only expected
to be used with test only module targets and base features.

Use the basic variant info to have access to application ids of all
build variants.

Bug: 256669321
Test: GradleApplicationIdProvider
Change-Id: I103bcbc72c2cc10a9c8a1dfdde5158fd101b03c2