Move ScrollPane into the DesignSurface of compose preview
Before this change, compose has a ScrollPanel contains the
DesignSurface. The size of DesignSurface may be changed when the size
or the number of previews are changed.

The scroll position cannot response to the change of SceneViewPanel
immediately because it needs to wait until the size of DesignSurface
is changed. If there is any painting event happens before responding,
the flicking happens.

Using the built-in ScrollPane in DesignSurface itself so the scroll
position can response to the relayout of SceneViewPanel immediately.

After this change, the adaptive scroll position change of vertical
list layout should be applied to compose preview.

Bug: 245363234
Test: All tests in compose-designer should pass
Change-Id: I262c1f61fcc6ce9267f140471740caa05b6fd7ee