Add memory Perfetto TraceConfig creation builder
Because native allocation tracking is a trace based
tool (uses perfetto), the Perfetto TraceConfig
proto can be extended for memory tracing usage on
top of the current cpu tracing. Thus, construction
of a memory based TraceConfig was created as a static
method to be used in startup and ui-triggered memory
tracing for when we replace StartNativeSample with
the unified TraceConfiguration proto.

The method(s) to construct cpu based Perfetto TraceConfig
have overlap with the memory based TraceConfig.
Therefore, the construction methods have been moved
out of PerfettoConfiguration (which is exclusively
for cpu tracing), and into the same file the memory
trace config builder methods are. Now all perfetto
based TraceConfig creation is localized.

This is part of a larger effort to unify the memory
tracing with cpu tracing. The end goal would be to
have native alloc tracking use a unified START_TRACE
command (along with cpu tracing), holding the unified
TraceConfiguration proto, instead of a separate command
and config proto. This change naturally prepares us for
thes forthcoming refactors.

Bug: 258515754
Test: added
Change-Id: I5935a46acf1d86878de071c4abb706b0ae672d73