Refresh VFS when Gradle build finishes
This is fixing an issue where AIDL sources and other generated files
are not being registered by Studio after building. This leads to
unresolved references to the AIDL sources in code. The fix causes
Studio to refresh when build finishes, and the sources are all
registered and references resolved after build.

Validated with a test that ensures that the scheduleRefresh() method is
called. Also validated with sample AIDL project
and tested that the build reflects the changes in Studio.

Bug: 251917804
Test: GradleSaveAndSyncTest
Change-Id: I50c238ac4a154a62a37c3f1e319f611009a14760