Create AndroidRunConfigurations with LAUNCH_DEFAULT_ACTIVITY by default
This fixes cases when a source manifest doesn't declare an activity, so hasDefaultLauncherActivity returns false, and we set launch options to DO_NOTHING.

When a user runs configuration we do nothing, it's confusing because we do it without any notification.

This fix restores behavior prior 5b5d5e5 for not-wear projects.

Bug: 249095314
Test: AndroidRunConfigurationsTest
Change-Id: Idf7d2ab1b94fd766c18a6117424a92ee7ba51a3e