Replace UserOptions in ProfilingConfiguration.fromProto
The ProfilingConfiguration::fromProto() method uses the
TraceConfiguration's UserOptions field to facililate the
conversion to a ProfilingConfiguration. This change replaces
the use of UserOptions with the technology-specific options
proto field in TraceConfiguration.

The fromProto message now does not include the config name
during the conversion. This is because the config name was
originally taken from UserOptions, but we are not using that
proto anymore. Because the profiler only uses the config name
to display in the dropdown before recording, the name is no
longer needed, allowing for this change.

Bug: 258515754
Test: existing
Change-Id: Id3db9dc776285b19feca6bc6fae40d76962a77cd