Maintain binding context & module descriptor connection.
The code used to discard the module descriptor from the
AnalysisResult pair and fetch it from the resolution facade
when needed. Unfortunately, that doesn't work because
the module descriptor is in a cache that can be flushed
in the resolution facade. Therefore, we could end up
passing mismatching module descriptor and binding context
into the compiler. This leads to situations where the
compiler has multiple descriptors for the same primitive
types. Multiple descriptors for the same primitive type
can lead to incorrect code generation.

Fixes: 256957527
Fixes: 244170328
Test: None. :( No idea how to test IDE cache flushes.
Change-Id: I17d3bd7daffd2f3dff0f32be6f090e3f7356d354