Remove unnecessary render from modelDerivedData
modelDerivedData is expected to happen after LayoutlibSceneManager has
completed the renders so additional data might be available.
This CL tries to remove the flag setRerenderWhenModelDerivedDataChanged
by making the behaviour as if it was always false.

Some code relied on modelDerivedData triggering a render but this mostly
caused unnecessary duplicated renders with no additional information.

Test: Existing tests pass
Bug: N/A
Change-Id: Id064bc0426ef5e7c45b9ac2f1a0e11097a946005