Remove UserOptions from legacy & startup profiling
In ag/20580476, the startup and legacy profiling start
capture calls have been updated to enable the use of
the unified trace config. This makes using UserOptions
obsolete. Thus, the removal of UserOptions was made in
any TraceConfiguration creation. There was no non-default
instance of TraceConfiguration for the legacy start profiling
call, but for startup tracing there was one instance.

Fixes: 259116828
Bug: 258515754
Test: manual
Change-Id: Ic67895a709f4905c47e3a8a5bedfb4000e9a5a2c