Remove TRACE_TYPE_MAP in TraceConfiguration creation
An earlier change (ag/20427704) introduced a temp
map to faciliate the removal of the proto enum TraceType
to the studio-side enum TraceType. Due to the changes that
followed, it is no longer needed. And actually, because of
the change done in ag/20544834, an error was caused by the
lack of new unified config proto options (tech-specific
options) in the TraceConfiguration constructed in
CpuCaptureStage's insertImportedTraceEvent method. Instead it
had the TRACE_TYPE_MAP-wrapping UserOptions proto.

Thus, this change removes the uncessary map uses and replaces
it with the new unified config proto where necessary. This
fixes the aforementioned error and cleans up the codebase.

It is also worth noting that not only was the TRACE_TYPE_MAP
removed, but also its wrapping UserOptions proto. The
UserOptions proto's only purpose was to wrap the trace type
and thus is also deemed unecessary and can be replaced by
a default instance of the trace-specific options.

After this change, only one instance of TRACE_TYPE_MAP is left
in the ProfilingConfiguration toProto method. This cannot be
removed until the toProto method itself and its usages are
removed (planned to be done in future changes).

Fixes: 258542374
Bug: 258515754
Test: manual
Change-Id: I537b2999e8b6cb05436260b2ef578cdabaca01fa