Configuration.toString method should not have side effects
Some of the getters used in the Configuration.toString method have side
effects, such as setting the value of the field it is getting when if it
wasn't previously set. That is an issue when debugging, as the debugger
will call the toString method and so set values that were not set by the
program itself.
Instead of calling getter methods, toString should simply look up the
value of fields directly.

Bug: N/A
Test: N/A
Change-Id: Ie843be98a77bdbf178ee29bda6b2d79b7eac7e30