[UI API] IDEA-307083, IDEA-306789 Change SimpleNode.doUpdate() signature
This is a follow-up to the recent changes in PresentableNodeDescriptor.
Now that the presentation is not published before it finishes being updated,
some existing workflows that relied on the old behavior are broken.

Notably, SimpleNode would call doUpdate() and expect it to change the
current presentation and/or some other presentation-related fields
such as myName and myIcon.

This, of course, had the same race condition as PresentableNodeDescriptor:
if all of these activities are performed on BGT (which is very often the case),
then there's a race condition whenever EDT decides to access them
(while painting, for example).

One of the SimpleNode subclasses affected was FilteringTreeStructure.FilteringNode,
which is the reason why the Structure Popup was broken (IDEA-306789).

The only viable solution is to prohibit altering the object's state in doUpdate().
However, then the state must be made available in some other way. Since there
is no external usages, the most client-friendly way is to simply provide the
presentation being updated as a parameter, but all implementations,
of course, had to be updated as well, which is why there are so many changed
files in this commit.

As for FilteringNode in particular, it would make most sense to simply override
update() instead of doUpdate(), as it's a typical delegating class, so why would
it invoke a lot of its own logic (inherited from SimpleNode.update)? Alas, simply
changing doUpdate() to update() would change behavior in an unpredictable
way, so for now we just leave that logic alone and instead fix the delegating
itself: instead of changing our own state, follow the same rule of updating
just the presentation (now passed as a parameter).

The next logical step would be to drop all that legacy nonsense from
SimpleNode.update() and PresentableNodeDescriptor.update()/apply(),
but, unfortunately, due to a large number of subclasses and clients,
it's not clear at the moment how it could be done without breaking a lot
of things.

(cherry picked from commit f7e3d155fe3d8a7e3035ab1ebacd21491691eb99)

Change-Id: If7e3d155fe3d8a7e3035ab1ebacd21491691eb99