Fix device dropdown menu keeps appearing when selecting nested device
When all the actions in the sub-menu of the popup menu are Toggleable,
the sub-menu becomes MultiChoice mode and the popup menu would not
disappear after the action is performed.

For the actions in the layout editor and compose preview, we always
hide the popup menu when a action is performed.

Fixes: 256565649
Test: manually (UI-related)
Change-Id: I3c951ebe7186b0754b8f8cfcb273fb46af433d3f
(cherry picked from commit aeac9ecb7fd80cd91f956ba1d7c3afc6b1cb5358)