[BA] ungroup actions as history feature is not used
BA history feature remains off in current release. This means
action used to access it is hidden which leaves Build Analyzer
menu actions group contain only single item and it does not look
good.
This change ungroup these actions so that when the second is hidden
there is no unnecessary grouping.
Also add a guard to BuildAttributionUiManager to make sure code
related to historical data will not run with flag disabled.

Bug: N/A
Test: Existing
Change-Id: I6b31b55baa2c173e2f52366a6b6345b9dd7e3e9d