Replace UserOptions in legacy cpu cmd handler
The LegacyCpuTraceCommandHandler utilizes UserOptions
in both the startTrace and stopTrace methods. As part
of the larger effort to remove UserOptions from the codebase,
this change replaces UserOptions with ArtOptions. This is
because at the each of the start and stop trace methods theres
an assertion that the TraceType is ART, so we can confidently
use the ArtOptions as a replacement.

Bug: 258515754
Test: existing
Change-Id: I415be2fb25e88223a29a2f1ac6d3d787033e359e