Fix Compose Preview toolbar not dismissing in edge cases
The Compose Preview toolbar is shown when hovering the toolbar area, and
dismissed when the mouse exits the area. In order to detect whether we
exited the area, we were calling JComponent#contains(Point), but this
method doesn't take into account the component's borders. This is now
fixed, as we're subtracting the border from the checked area.

Another edge case fixed by this change is when the current window is put
on background (e.g. through an alt + tab command), which would trigger a
mouseExit event, but since the mouse coordinates could still be on top
of the toolbar, we would fail to hide it. Now we're always hiding the
toolbar when the current window is not focused.

Fixes: 260996054
Test: N/A. Visual behavior.
Change-Id: I13cac978facf9cee5cbcaf81b5523f51f7bd4d81
(cherry picked from commit 450febaea21afd380c00366605fb2f20691798e1)