Refactor annotated search into strategy object
This is step 1 of a change to allow a switch between using the custom
Dagger annotation search and using the built-in annotation search. This
commit just moves around code, but does not change any functionality.
(And this is verified by the unit tests, which have good coverage and
don't see any differences in logic.)

Bug: 255299583
Test: DaggerUtilTest
Change-Id: Ic1ccddbc28253de4d7a0a84effa50c0b92540492