Add build in search strategy behind flag
This change adds the ability to switch to the built-in annotations
search behind a studio flag, allowing easy experimentation between the
two variants. This is effectively a rollback of ag/12009170, albeit with
a slightly different structure.

The only new code in DaggerUtil is for the
getAssistedFactoryMethodsForType method, which didn't exist at the time
of ag/12009170. Everything else is restoring old logic that used to be
there. This is verified by the unit tests, which are running and passing
with both search variants.

Bug: 255299583
Test: DaggerUtilTest
Change-Id: I3d168b5372244786e8fbd26b201829929facbeef