[NPW] Use existing version of Kotlin if possible
In an ideal world, if a project is already using Kotlin, the new
module wizard should not need to make modifications to existing build
configuration files (in particular those shared by other Gradle
projects such as the root build or settings files).

We are not in an ideal world.  At the moment, it is entirely possible
to declare use of a Kotlin Gradle plugin plugin in a leaf
module (through the new-form "plugins Dsl" with its infix `version`
operator, either directly or implicitly through use of Version
Catalogs).  The New Module apparatus does not support this idiom, and
so consequently adds a plugin declaration in the root build.gradle
file, and uses it in the newly-added module.

From the perspective of any existing modules using a Kotlin plugin,
this plugin declaration in the root is redundant, except that if the
version declared in the root conflicts with the version declared in
the existing module, gradle sync will fail.  So we should try harder
to add declarations, if we're adding them, which do not conflict with
existing uses: which for our purposes means that if all source sets in
modules in all linked projects use the same version of the Kotlin
plugin, we should use that instead of our existing default.  (If it is
not the case that all source sets in all modules in all linked
projects use the same version of the Kotlin plugin, then we're
probably doomed.)

Bug: 259512836
Test: manual
Change-Id: Ia78a8f324b0181e95337321963fa4b0720d630e7